move capturing of screenshots to ImageCapture utility
Extracts the logic for caturing/saving/uploading
images from the screenshot plugin to a separate
utility named ImageCapture.
The Imgur class were made inner classes of ImageCapture
since the only location they see use is there.
ImageUploadStyle and TransferableBufferedImage were
also moved to ulities to allow ImageCapture and other
plugins to use them.
