+++ runelite-client/src/main/resources/net/runelite/client/runelite.properties
@@ -13,3 +13,4 @@ runelite.jav_config=http://oldschool.runescape.com/jav_config.ws
 runelite.jav_config_backup=http://static.runelite.net/jav_config.ws
 runelite.pluginhub.url=https://repo.runelite.net/plugins
 runelite.pluginhub.version=${project.version}
+runelite.imgur.client.id=30d71e5f6860809
